Improved VisualRender memoization (#222)
Lots of surrounding changes to additional components, to prevent Vega from re-rendering unnecessarily.

This will correctly memoize the Vega view for init and update of the visual in standard view, and the open/close of the editor and dialogs.

There may be further dependencies to explore throughout the editor, but this is a big change and would like to ensure that i's safe before proceeding further.