Supply cloned objects to react-vega
By introducing memoization in #222, when we do a comparison on spec/config/data, we didn't take into account that react-vega may mutate them, e.g. transforms applied during processing. This was causing an infinite render loop and killing the main window.

This resolves #248 by cloning the objects first before suppling them to react-vega (like we were originally doing). I've added a test case for this to my Power BI workbook also, so we can check it for any subsequent major changes to the `VisualRender` component in future.