perf!: move Vega view from store to singleton and consolidate code
Also moves rendering events to correct place in the render lifecycle, based on when the view is instantiated, which will hopefully mitigate some issues with copy as image.

BREAKING CHANGE: Signal and dataset viewers need to be re-written to use the view API correctly. These have been replaced with a placeholder element and will be fixed in a subsequent build as part of this release.