fix: revise hashing approach and sequencing (#357)
Have profiled the visual when processing the spec and dataset, and our initial approach of merging the data into the spec at parse time is too heavy on general performance.

We now:

- Parse the spec with the patched properties (except data).
- Hash the dataset prior to generating selection IDs.
- Use a more performant hashing algorithm, as we only need a unique ID and not actual encryption.
- Centralized the hashing function so that we use this wherever we need to hash things.

Improvements are significant for visuals with very large datasets. The S&P example now initializes in-line with 1.5 (the majority of which is essentially the cost of querying the data model from Power BI). We've also been able to improve the transitions to and from the visual and editor even further.